Generic autobumper can optionally skip running when there is no oncall
This can be used to automatically stop automated deployment of prow during long weekends
